Adds more entities controllers daos services

+++ dancewear/src/main/java/com/example/dancewear/dao/ReviewsRepository.java
@@ -3,5 +3,5 @@ package com.example.dancewear.dao;
 import com.example.dancewear.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewsRepository extends JpaRepository<Review, Long> {
+public interface ReviewsRepository extends JpaRepository<Review, Integer> {
 }
